% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hai_totals.R
\name{plot_hai_totals}
\alias{plot_hai_totals}
\title{Plot total HAI counts}
\usage{
plot_hai_totals(data, by = "Infection_Type")
}
\arguments{
\item{data}{Data frame (like \code{hai_data_clean})}

\item{by}{Grouping variable ("Infection_Type", "Age_Group", or "Sex")}
}
\description{
Creates a bar chart of total infections by a chosen variable.
}
