% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hai_totals.R
\name{plot_hai_totals}
\alias{plot_hai_totals}
\title{Plot total HAI counts}
\usage{
plot_hai_totals(data, by = "Infection_Type", show_prop = FALSE, pal = NULL)
}
\arguments{
\item{data}{Data frame like \code{hai_data_clean}}

\item{by}{One of "Infection_Type", "Age_Group", or "Sex"}
}
\description{
Creates a bar chart of total infections by a chosen variable.
}
\examples{
# data("hai_data_clean", package = "haiInsight")
# plot_hai_totals(hai_data_clean, by = "Infection_Type")
# plot_hai_totals(hai_data_clean, by = "Age_Group")
# plot_hai_totals(hai_data_clean, by = "Sex")

}
